lib/actions/dom-actions/is-disabled/is-disabled.test.ts
import * as puppeteer from 'puppeteer-core';
import * as SUT from '../index';
import { launchBrowser } from '../../browser-actions';
import { getChromePath } from '../../../utils';
import * as path from 'path';

describe('selector is disabled', (): void => {
  let browser: puppeteer.Browser | undefined = undefined;
  beforeEach((): void => {
    jest.setTimeout(30000);
  });
  afterEach(
    async (): Promise<void> => {
      if (browser) {
        await browser.close();
      }
    },
  );
  test('should return an error when page has not been initalized', async (): Promise<void> => {
    // Given
    const page: puppeteer.Page | undefined = undefined;
    // When
    // Then
    const expectedError = new Error(
      "Error: cannot get the disabled property of 'foobar' because a new page has not been created",
    );
    await SUT.isDisabled('foobar', page).catch((error): void =>
      expect(error).toMatchObject(expectedError),
    );
  });

  test('should return an error when selector is not found', async (): Promise<void> => {
    // Given
    browser = await launchBrowser({
      headless: true,
      executablePath: getChromePath(),
    });
    const page = await browser.newPage();

    // When

    // Then
    const expectedError = new Error('Error: failed to find element matching selector "foobar"');
    await SUT.isDisabled('foobar', page).catch((error): void =>
      expect(error).toMatchObject(expectedError),
    );
  });

  test('should detect that input selector is disabled', async (): Promise<void> => {
    // Given
    browser = await launchBrowser({
      headless: true,
      executablePath: getChromePath(),
    });
    const page = await browser.newPage();
    await page.goto(`file:${path.join(__dirname, 'is-disabled.test.html')}`);

    // When
    const isDisabled = await SUT.isDisabled('#disabledInput', page);

    // Then
    expect(isDisabled).toBe(true);
  });

  test('should detect that select selector is disabled', async (): Promise<void> => {
    // Given
    browser = await launchBrowser({
      headless: true,
      executablePath: getChromePath(),
    });
    const page = await browser.newPage();
    await page.goto(`file:${path.join(__dirname, 'is-disabled.test.html')}`);

    // When
    const isDisabled = await SUT.isDisabled('#disabledSelect', page);

    // Then
    expect(isDisabled).toBe(true);
  });

  test('should detect that input selector is enabled', async (): Promise<void> => {
    // Given
    browser = await launchBrowser({
      headless: true,
      executablePath: getChromePath(),
    });
    const page = await browser.newPage();
    await page.goto(`file:${path.join(__dirname, 'is-disabled.test.html')}`);

    // When
    const isDisabled = await SUT.isDisabled('#enabledInput', page);

    // Then
    expect(isDisabled).toBe(false);
  });
});
